move EventSourcing.Serialization.Json DI code into main EventSourcing.Serialization.Json package
