+++ EventSourcing.Bus/Bus.RabbitMQ.IntegrationTests/Bus.RabbitMQ.IntegrationTests.csproj
@@ -24,7 +24,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\Core\EventSourcing\EventSourcing.csproj" />
-      <ProjectReference Include="..\..\EventSourcing.Serialization\Extensions.DependencyInjection.Serialization.Json\Extensions.DependencyInjection.Serialization.Json.csproj" />
+      <ProjectReference Include="..\..\EventSourcing.Serialization\Serialization.Json\Serialization.Json.csproj" />
       <ProjectReference Include="..\..\Helpers\TestHelpers\TestHelpers.csproj" />
       <ProjectReference Include="..\Bus.RabbitMQ\Bus.RabbitMQ.csproj" />
       <ProjectReference Include="..\Bus\Bus.csproj" />
