+++ EventSourcing.Bus/Bus.RabbitMQ.IntegrationTests/RabbitMQCollectionFixture.cs
@@ -4,7 +4,7 @@ using EventSourcing.Abstractions.Handling;
 using EventSourcing.Bus.Extensions;
 using EventSourcing.Bus.RabbitMQ.Extensions;
 using EventSourcing.Extensions;
-using EventSourcing.Extensions.DependencyInjection.Serialization.Json;
+using EventSourcing.Serialization.Json.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
