+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/ForgettablePayloads.Persistence.PostgreSql.IntegrationTests/PostgreSqlCollectionFixture.cs
@@ -4,11 +4,11 @@ using System.Threading;
 using System.Threading.Tasks;
 using DbUp;
 using EventSourcing.Extensions;
-using EventSourcing.Extensions.DependencyInjection.Serialization.Json;
 using EventSourcing.ForgettablePayloads.Extensions;
 using EventSourcing.ForgettablePayloads.Extensions.DatabaseMigrations.Persistence.PostgreSql.DbUp.Extensions;
 using EventSourcing.ForgettablePayloads.Persistence.PostgreSql;
 using EventSourcing.ForgettablePayloads.Persistence.PostgreSql.Extensions;
+using EventSourcing.Serialization.Json.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
