+++ EventSourcing.ForgettablePayloads/Persistence/SqlServer/ForgettablePayloads.Persistence.SqlServer.IntegrationTests/SqlServerCollectionFixture.cs
@@ -3,11 +3,11 @@ using System.Data.SqlClient;
 using System.Threading.Tasks;
 using DbUp;
 using EventSourcing.Extensions;
-using EventSourcing.Extensions.DependencyInjection.Serialization.Json;
 using EventSourcing.ForgettablePayloads.Extensions;
 using EventSourcing.ForgettablePayloads.Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp.Extensions;
 using EventSourcing.ForgettablePayloads.Persistence.SqlServer;
 using EventSourcing.ForgettablePayloads.Persistence.SqlServer.Extensions;
+using EventSourcing.Serialization.Json.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
