+++ EventSourcing.Serialization/Serialization.Json.UnitTests/Serialization.Json.UnitTests.csproj
@@ -7,6 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.1" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
         <PackageReference Include="xunit" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
@@ -20,6 +21,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\Core\EventSourcing\EventSourcing.csproj" />
       <ProjectReference Include="..\..\Helpers\TestHelpers\TestHelpers.csproj" />
       <ProjectReference Include="..\Serialization.Json\Serialization.Json.csproj" />
     </ItemGroup>
