+++ EventSourcing.Serialization/Serialization.Json/Serialization.Json.csproj
@@ -28,6 +28,7 @@
     <ItemGroup>
       <ProjectReference Include="..\..\Core\Abstractions\Abstractions.csproj" />
       <ProjectReference Include="..\Serialization.Abstractions\Serialization.Abstractions.csproj" />
+      <ProjectReference Include="..\Serialization\Serialization.csproj" />
     </ItemGroup>
     <ItemGroup>
       <PackageReference Include="System.Text.Json" Version="6.0.6" />
